Remove unnecessary intermediate file
